user opetations with admin authorization

+++ routes/user.py
@@ -1,11 +1,11 @@
-from fastapi import APIRouter, Body
+from fastapi import APIRouter, Body,Header
 from fastapi.encoders import jsonable_encoder
 from passlib.context import CryptContext
 from fastapi.security import HTTPBasicCredentials
 
 from database.user_database import *
 from models.user import *
-from auth.jwt_handler import signJWT
+from auth.jwt_handler import signJWT, decodeJWT
 
 user_router = APIRouter()
 user_login_router = APIRouter()
@@ -40,34 +40,49 @@ async def user_login(user_credentials: HTTPBasicCredentials = Body(...)):
 
 
 @user_router.get("/", response_description="Users retrieved")
-async def get_users():
-    users = await retrieve_users()
-    return ResponseModel(users, "Users data retrieved successfully") \
-        if len(users) > 0 \
-        else ResponseModel(
-        users, "Empty list returned")
+async def get_users(authorization:Optional[str]=Header(None)):
+    token_data = decodeJWT(authorization.split(' ')[1])
+    is_admin = (await retrieve_user(email=token_data['user_id']))['is_admin']
+    if is_admin:
+        users = await retrieve_users()
+        return ResponseModel(users, "Users data retrieved successfully") \
+            if len(users) > 0 \
+            else ResponseModel(
+            users, "Empty list returned")
+    return "Permission denied"
 
 
 @user_router.get("/{id}", response_description="User data retrieved")
-async def get_user_data(id):
-    user = await retrieve_user(id)
-    return ResponseModel(user, "User data retrieved successfully") \
-        if user \
-        else ErrorResponseModel("An error occured.", 404, "User doesn't exist.")
+async def get_user_data(id: str, authorization:Optional[str]=Header(None)):
+    token_data = decodeJWT(authorization.split(' ')[1])
+    is_admin = (await retrieve_user(email=token_data['user_id']))['is_admin']
+    if is_admin:
+        user = await retrieve_user(id)
+        return ResponseModel(user, "User data retrieved successfully") \
+            if user \
+            else ErrorResponseModel("An error occured.", 404, "User doesn't exist.")
+    return "Permission denied"
 
 
 @user_router.delete("/{id}", response_description="User data deleted from the database")
-async def delete_user_data(id: str):
-    deleted_user = await delete_user(id)
-    return ResponseModel("User with ID: {} removed".format(id), "User deleted successfully") \
-        if deleted_user \
-        else ErrorResponseModel("An error occured", 404, "User with id {0} doesn't exist".format(id))
-
+async def delete_user_data(id: str, authorization:Optional[str]=Header(None)):
+    token_data = decodeJWT(authorization.split(' ')[1])
+    is_admin = (await retrieve_user(email=token_data['user_id']))['is_admin']
+    if is_admin:
+        deleted_user = await delete_user(id)
+        return ResponseModel("User with ID: {} removed".format(id), "User deleted successfully") \
+            if deleted_user \
+            else ErrorResponseModel("An error occured", 404, "User with id {0} doesn't exist".format(id))
+    return "Permission denied"
 
 @user_router.put("{id}")
-async def update_user(id: str, req: UpdateUserModel = Body(...)):
-    updated_user = await update_user_data(id, req.dict())
-    return ResponseModel("User with ID: {} name update is successful".format(id),
-                         "User name updated successfully") \
-        if updated_user \
-        else ErrorResponseModel("An error occurred", 404, "There was an error updating the user.".format(id))
+async def update_user(id: str, req: UpdateUserModel = Body(...), authorization:Optional[str]=Header(None)):
+    token_data = decodeJWT(authorization.split(' ')[1])
+    is_admin = (await retrieve_user(email=token_data['user_id']))['is_admin']
+    if is_admin:
+        updated_user = await update_user_data(id, req.dict())
+        return ResponseModel("User with ID: {} name update is successful".format(id),
+                            "User name updated successfully") \
+            if updated_user \
+            else ErrorResponseModel("An error occurred", 404, "There was an error updating the user.".format(id))
+    return "Permission denied"
